refactor(attribute parser): move check_custom_mir to attribute parser
